requirements.txt
streamlit
openai
email
pdfplumber
PyMuPDF
pymongo
python-docx
langchain
python-dotenv